v2015121500 add AllListsViewController
Two important views:
1.About implicitly unwrapped optionals like
var checklist: Checklist!
That allows its value to be temporarily nil until viewDidLoad()
happens.nil is normally not an allowed value for variables in Swift but
by using the ! you override that.
2. By writing “sender as! some” you tell Swift that it can safely treat
sender as some object.Note that there is also as ? .This is for casting
optionals,or when the type cast is allowed to fail